requirements.txt
paho-mqtt
jinja2
python-multipart
Flask
gunicorn
Flask-SQLAlchemy
